Gather and display useful data.
I'm interested in displaying a dashboard showing the current state of
pull requests in an org, with maybe some historical context. The key
variables I'm interested in are duration (time spent in review) and
depth (number of PRs outstanding at a given time).

There are two options for getting the data: (1) indexing it by
spidering through all the pages or (2) getting the most recent PRs
up to an arbitrary limit- like the first page. I've decided to go
with 2 because its much less complex and the default of 30 per page
should get plenty of historical data for most of the projects I work
with. This commit implements option 2 and dumps out both opened and
closed PRs along with the key variables.