Handle pull requests from multiple repositories.
I realized that a dashboard is mostly useful for displaying current
actionable information (for example, a line graph is useful for
seeing anomalies like a spike in traffic) and closed PRs aren't
actionable. We want 0 WIP PRs so any open PR is actionable.

This commit removes the check on closed PRs and adds support for
multiple repositories.