Support filtering by author.
In organizations with monolithic applications, there might be a ton
of PRs where only a few are relevant to a project team (specifically,
other team member's PRs). This commit allows depth to be controlled
on a per repo basis and specific authors to be monitored. If the
author list isn't given, we'll just show everything.