Deployment script.
I'm not a fan of committing third-party dependencies to github. The
heroku godep integration requires them to be included or the build
will fail. This commit automates a process to tack them onto the
heroku branches and then deploys the branch as master. However, the
actual master branch does not have vendor/ committed.

This is a compromise solution (I'd rather heroku just did some type
of go get). If this wasn't a public repository, I'd be very tempted
to just include the vendor/ directory for simplicity.