setup feedback loop for front-end dev
One of the standards I've been trying out recently is that you should
be able to change the code without running application in a live
environment (locally, staging, etc.) and not break production. This
is important right now because SSLRedirect makes local testing a bit
unpleasant.

To address this issue, this commit adds a test and some new render
code that generates the dashboard assets for sample data in a temp
file that can be opened in a browser. This lets me iterate on the
rendering code without having to actually run the full application
(fast feedback!). Note that I haven't migrated the existing front-end
code yet.