only render one pr block
My track record for deploying without running a local instance
has broken the experience on production a few too many times,
but hopefully will get better as the tests improve.

This commit fixes a bug where the Render function was called within
a for loop, causing too many blocks to be shown.