Begin to place dashboard under test.
In order to deal with my habit of breaking production, this commit
starts to add very broad end-to-end testing around the dashboard.
Go-vcr is used to fetch down sample PRs and then the output is dumped.
The next step is to assert the output equals a recorded output so that
I can safely do some refactoring (actually, this is just for fun).

The main problem is controlling time, which changes on each test run.