Use channels for pipeline.
The original dashboard function retrieved and assembled the data in
one go- which is hard to place under unit coverage. Under the cover
of the dashboard regression test, this commit refactors the function
to use channels and go routines as a pipeline. This may not be the
right approach(?), but its a good opportunity to get better at
channels.