Reduce refresh rate to once every 24 hours.
The free heroku dyno is kept running constantly on a 10 minute refresh
cycle. From watching usage over the last week, I've noticed that the
primary value was the "accountability" the monitor creates- frequency
of updates is less important. Its easier to check github to see which
PRs need to be reviewed, but the dashboard creates an atmosphere.

This commit refreshes the dashboard to keep the atmosphere but spare
the heroku dyno.