Begin building out gantt chart.
This commit prototypes a gantt chart. A major problem with the
original "snapshot" model is that you can't measure team progress
and improvements over time- instead of being able to get the morale
boost from "we're going faster" you just see a big mountain of work
to accomplish. Using a gantt chart lets a team see the historical
depths and durations so progress becomes more visible and hopefully
rewarding.