Filter all repos by optional author list.
This commit adds back author filtering, but applies it to the
entire set of repositories. This dashboard is intended to be used
by specific teams that all review eachothers PRs, so this type of
filtering makes a lot of sense (especially when the team is working
in a large codebase with many other teams).