Use merge to handle parallel retrieval.
This commit makes retrieve more general by having it just pull one
type of PR, then calling retrieve multiple times. The output channels
are merged into the rest of the pipeline. The advantage to this
approach, besides a simpler retrieve, is the ability to parallelize
retrieve so we aren't blocking wating for http responses.