Add grid rules to gantt chart.
While the vertical rules add some visual clutter, I think it is an
important affordance to the user- the rules orient them in time and
the meaning of the bars more obvious. Or so I hope. This commit also
adds a visual guide to reading the charts to the README.